# docs/swagger.yaml
basePath: /api
definitions:
  models.Airport:
    properties:
      eventId:
        type: integer
      id:
        type: integer
      identifier:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      maximumAircraftPerHour:
        type: integer
      numberOfVotes:
        type: integer
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.Location:
    properties:
      eventId:
        type: integer
      id:
        type: integer
      identifier:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      maximumAircraftPerHour:
        type: integer
      routeSegmentId:
        type: integer
      sortOrder:
        type: integer
    type: object
...
  models.RouteSegment:
    properties:
      color:
        type: string
      enabled:
        type: boolean
      eventId:
        type: integer
      facilities:
        type: string
      id:
        type: integer
      identifier:
        type: string
      locations:
        items:
          $ref: '#/definitions/models.Location'
        type: array
      maximumAircraftPerHour:
        type: integer
      providedFacilityProgression:
        items:
          $ref: '#/definitions/models.Sector'
        type: array
      routeRevision:
        type: integer
      routeSegmentGroup:
        type: string
      routeString:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.RouteSegmentTag'
        type: array
    type: object
  models.RouteSegmentTag:
    properties:
      id:
        type: integer
      routeSegmentId:
        type: integer
      tag:
        type: string
    type: object
  models.Sector:
    properties:
      eventId:
        type: integer
      id:
        type: integer
      identifier:
        type: string
      maximumAircraftPerHour:
        type: integer
      sectorBoundaries:
        items:
          $ref: '#/definitions/models.SectorBoundary'
        type: array
    type: object
  models.SectorBoundary:
    properties:
      coordinates:
        items:
          $ref: '#/definitions/models.SectorBoundaryCoordinate'
        type: array
      id:
        type: integer
      maxLatitude:
        type: number
      maxLongitude:
        type: number
      minLatitude:
        type: number
      minLongitude:
        type: number
      sectorId:
        type: integer
    type: object
  models.SectorBoundaryCoordinate:
    properties:
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      sectorBoundaryId:
        type: integer
      sortOrder:
        type: integer
    type: object
...
  models.Slot:
    properties:
      arrivalAirport:
        $ref: '#/definitions/models.Airport'
      arrivalAirportId:
        type: integer
      departureAirport:
        $ref: '#/definitions/models.Airport'
      departureAirportId:
        type: integer
      departureTime:
        type: string
      id:
        type: integer
      projectedArrivalTime:
        type: string
      routeSegments:
        items:
          $ref: '#/definitions/models.RouteSegment'
        type: array
      slotRevisionId:
        type: integer
    type: object
...
  models.SlotRevision:
    properties:
      createdAt:
        type: string
      eventId:
        type: integer
      id:
        type: integer
      number:
        type: integer
      simulationOutputCommentary:
        type: string
      slotGenerationOutputCommentary:
        type: string
      slots:
        items:
          $ref: '#/definitions/models.Slot'
        type: array
      throughputSnapshots:
        items:
          $ref: '#/definitions/models.ThroughputSnapshot'
        type: array
      throughputStates:
        items:
          $ref: '#/definitions/models.ThroughputState'
        type: array
    type: object
  models.SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
  models.ThroughputSnapshot:
    properties:
      id:
        type: integer
      minuteOffset:
        type: integer
      slot:
        $ref: '#/definitions/models.Slot'
      slotId:
        type: integer
      slotRevisionId:
        type: integer
      throughputPointId:
        type: integer
      throughputPointType:
        type: string
    type: object
  models.ThroughputState:
    properties:
      departureTimeWindowStart:
        type: string
      id:
        type: integer
      maximumSlots:
        type: integer
      slotRevisionId:
        type: integer
      slotsAllocated:
        type: integer
      throughputPointId:
        type: integer
      throughputPointType:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: Central data API for CTP (Cross the Pond) event management. All /api/*
    endpoints require an X-API-Key header.
  title: CTP API
  version: 1.0.0
paths:
...
  /events/{eventId}/slot-revisions:
    get:
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SlotRevision'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List slot revisions for an event
      tags:
      - slot-revisions
    post:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      - description: Slot revision (number auto-assigned if 0)
        in: body
        name: revision
        required: true
        schema:
          $ref: '#/definitions/models.SlotRevision'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SlotRevision'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create slot revision
      tags:
      - slot-revisions
  /events/{eventId}/slot-revisions/{number}:
    get:
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      - description: Revision number
        in: path
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SlotRevision'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get slot revision by number
      tags:
      - slot-revisions
  /events/{eventId}/slot-revisions/latest:
    get:
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SlotRevision'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get latest slot revision for an event (fully preloaded)
      tags:
      - slot-revisions
...
  /slot-revisions/{revisionId}:
    delete:
      parameters:
      - description: Slot revision ID
        in: path
        name: revisionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete slot revision
      tags:
      - slot-revisions
    put:
      consumes:
      - application/json
      parameters:
      - description: Slot revision ID
        in: path
        name: revisionId
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: revision
        required: true
        schema:
          $ref: '#/definitions/models.SlotRevision'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SlotRevision'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update slot revision metadata
      tags:
      - slot-revisions
  /slot-revisions/{revisionId}/slots:
    post:
      consumes:
      - application/json
      parameters:
      - description: Slot revision ID
        in: path
        name: revisionId
        required: true
        type: integer
      - description: Slots to add
        in: body
        name: slots
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Slot'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.Slot'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add slots to a slot revision
      tags:
      - slot-revisions
  /slot-revisions/{revisionId}/throughput-snapshots:
    post:
      consumes:
      - application/json
      parameters:
      - description: Slot revision ID
        in: path
        name: revisionId
        required: true
        type: integer
      - description: Throughput snapshots
        in: body
        name: snapshots
        required: true
        schema:
          items:
            $ref: '#/definitions/models.ThroughputSnapshot'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.ThroughputSnapshot'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add throughput snapshots to a slot revision
      tags:
      - slot-revisions
  /slot-revisions/{revisionId}/throughput-states:
    post:
      consumes:
      - application/json
      parameters:
      - description: Slot revision ID
        in: path
        name: revisionId
        required: true
        type: integer
      - description: Throughput states
        in: body
        name: states
        required: true
        schema:
          items:
            $ref: '#/definitions/models.ThroughputState'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.ThroughputState'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add throughput states to a slot revision
      tags:
      - slot-revisions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"